tg-subscription/requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
redis[hiredis]==5.1.0
celery[redis]==5.4.0
pydantic-settings==2.5.2
httpx==0.27.2

# 第三方支付
wechatpayv3==0.2.8
alipay-sdk-python-new==3.6.905
qrcode[pil]==7.4.2
